move the csv reader aroun, place in game-page as mutliple children'd like to resuse this value
