+++ src/pages/game-page.js
@@ -3,9 +3,16 @@ import React, { useState } from 'react';
 import { useParams } from 'react-router-dom';
 import Navbar from '../components/navbar';
 import AnswerCard from '../components/answer-card';
+import CsvReader from '../shared/csv-reader';
 
 function GamePage() {
 	const { gameKey } = useParams();
+	const databaseDict = JSON.parse(localStorage.getItem("databaseDict"));
+	const answerFilePath = "/csv/answers/"+databaseDict[gameKey]?.AnswerFile;
+	const [csvData, setCsvData] = useState([]);
+	const handleCsvData = (data) => {
+		setCsvData(data);
+	};
 
 	const [showPopup, setShowPopup] = useState(false);
 	const togglePopup = () => {
@@ -15,12 +22,16 @@ function GamePage() {
     return (
       	<div>
           	<Navbar title="Game" />
+			<CsvReader 
+                filePath={answerFilePath} 
+                onDataParsed={handleCsvData}
+                asDict={false} />
 			<button className="answer-popup-btn" onClick={togglePopup}>Answer</button>
       	
 			{showPopup && (
 				<div className="popup-overlay">
 					<div className="popup-box">
-						<AnswerCard inPopup index={gameKey} />
+						<AnswerCard inPopup index={gameKey} csvData={csvData}/>
 						<button className="answer-popup-btn" onClick={togglePopup}>Close</button>
 					</div>
 				</div>
